Improved the way transparency works. Removed reduntant code in the GUI system. Moved stuff from the Window class to the Desktop class. Text drawing respects the current transparency.

+++ roengine/gui/buttonbar.cc
@@ -103,7 +103,7 @@ void ButtonBar::Draw(unsigned int delay) {
 	}
 
 	// Right edge
-	Window(w - tex_right.getWidth(), y, (float)tex_right.getWidth(), (float)tex_right.getHeight(), tex_right);
+	Window((float)w - tex_right.getWidth(), y, (float)tex_right.getWidth(), (float)tex_right.getHeight(), tex_right);
 
 	// Position buttons
 	used_x = 0;
